Refactor: file the onboarding tour the way the rest of the page is filed
The tour was the one feature under pages/stacks that kept its own folder
instead of the components/hooks/lib split its siblings use, so finding any
part of it meant knowing it was special. The dialog, the hook and the tour
engine now sit with the other dialogs, hooks and engines.
