+++ frontend/src/pages/stacks/components/list/index.tsx
@@ -4,9 +4,9 @@ import { Navigate, useNavigate, useSearchParams } from "react-router-dom";
 import { useEffect, useMemo, useRef, useState } from "react";
 import { getStacksByOrg, deleteStack } from "@/api/stacks";
 import { getOrganization } from "@/api/organizations";
-import { buildHelloStackSeed } from "@/pages/stacks/onboarding/hello-stack-seed";
-import { startCanvasStage, isTourDone, markTourDone } from "@/pages/stacks/onboarding/tour";
-import { WelcomeDialog } from "@/pages/stacks/onboarding/welcome-dialog";
+import { buildHelloStackSeed } from "@/pages/stacks/lib/onboarding/hello-stack-seed";
+import { startCanvasStage, isTourDone, markTourDone } from "@/pages/stacks/lib/onboarding/tour";
+import { WelcomeDialog } from "@/pages/stacks/components/onboarding/welcome-dialog";
 import { useToast } from "@/components/ui/use-toast";
 import { useConfirm } from "@/components/branded/confirm";
 import { useResourceProjects } from "@/hooks/use-resource-projects";
