+++ frontend/src/pages/stacks/components/onboarding/welcome-dialog.tsx
