+++ frontend/src/pages/stacks/lib/onboarding/hello-stack-seed.ts
