+++ frontend/src/pages/stacks/lib/onboarding/tour.ts
